minor changes while testing, will remove a class
im going to remove the merger class and move the sort method so it is a
public method within the main class. no need to create a class to
instantiate once, ya know? 
also, theres a huge block of comments in the Mergesort class for you to
read. im trying to make sure we both have a decent understanding of the
indexing thats happening.